requirements.txt
gym==0.10.9
numpy>=1.16.1
yacs>=0.1.5
numpy-quaternion>=2019.3.18.14.33.20
attrs>=19.1.0
opencv-python>=3.3.0
cmake
pytest
pyquaternion
faiss-gpu
# visualization optional dependencies
imageio>=2.2.0
imageio-ffmpeg>=0.2.0
scipy>=1.0.0
tqdm>=4.0.0
numba>=0.44.0
